+++ tests/testthat/test-lint_free.R
@@ -1,11 +1,16 @@
+# Settings:
+skip_lintr <- FALSE
 
 if (requireNamespace("lintr", quietly = TRUE)) {
   test_that(
-    "Package Style",
+    desc = "[lintr] Package Style",
     code = {
-      lintr::expect_lint_free(
-        relative_path = FALSE
-      )
+      if (skip_lintr) {
+        skip("[skip_lintr: TRUE] lintr deactivated in test-lint_free.R")
+      } else {
+        invisible()
+      }
+      lintr::expect_lint_free(relative_path = FALSE)
     }
   )
 }